requirements.txt
python-dotenv==0.19.2
schedule==1.1.0
tweepy==4.5.0
